moved a bunch of stuff around for task queues
