+++ server/config.py
@@ -1,5 +1,5 @@
 import os
 #TEAM_NAME = os.environ['TEAM_NAME']
-SQLALCHEMY_DATABASE_URI = "postgresql+psycopg2:///mentionscrawler"
+SQLALCHEMY_DATABASE_URI = "postgresql+psycopg2:///mentions_crawler_flask"
 SQLALCHEMY_TRACK_MODIFICATIONS = False
 SECRET_KEY = "nOtSoSeCrEtKeY!!!"
